+++ rom/intuition/printitext.c
@@ -21,8 +21,8 @@
 /*  SYNOPSIS */
         AROS_LHA(struct RastPort  *, rp, A0),
         AROS_LHA(struct IntuiText *, iText, A1),
-        AROS_LHA(WORD              , leftOffset, D0),
-        AROS_LHA(WORD              , topOffset, D1),
+        AROS_LHA(LONG              , leftOffset, D0),
+        AROS_LHA(LONG              , topOffset, D1),
 
 /*  LOCATION */
         struct IntuitionBase *, IntuitionBase, 36, Intuition)
@@ -61,7 +61,7 @@
 } /* PrintIText */
 
 void int_PrintIText(struct RastPort * rp, struct IntuiText * iText,
-            	    WORD leftOffset, WORD topOffset, BOOL ignore_attributes,
+            	    LONG leftOffset, LONG topOffset, BOOL ignore_attributes,
 		    struct IntuitionBase *IntuitionBase)
 {
